+++ src/support/utils.js
@@ -136,12 +136,14 @@ export const sendAutofixMessage = async (
   opportunityId,
   suggestionIds,
   promiseToken,
+  variation,
   { url } = {},
 ) => sqs.sendMessage(queueUrl, {
   opportunityId,
   siteId,
   suggestionIds,
   promiseToken,
+  variation,
   url,
 });
 /* c8 ignore end */
@@ -701,12 +703,17 @@ export const onboardSingleSite = async (
     }
 
     // Resolve canonical URL for the site from the base URL
-    const resolvedUrl = await resolveCanonicalUrl(baseURL);
-    const { pathname: baseUrlPathName } = new URL(baseURL);
+    let resolvedUrl = await resolveCanonicalUrl(baseURL);
+    if (resolvedUrl === null) {
+      log.warn(`Unable to resolve canonical URL for site ${siteID}, using base URL: ${baseURL}`);
+      resolvedUrl = baseURL;
+    }
+    const { pathname: baseUrlPathName, origin: baseUrlOrigin } = new URL(baseURL);
+    log.info(`Base url: ${baseURL} -> Resolved url: ${resolvedUrl} for site ${siteID}`);
     const { pathname: resolvedUrlPathName, origin: resolvedUrlOrigin } = new URL(resolvedUrl);
 
-    // Update the fetch configuration only if the pathname is different from the resolved URL
-    if (baseUrlPathName !== resolvedUrlPathName) {
+    // Update the fetch configuration only if the pathname/origin is different from the resolved URL
+    if (baseUrlPathName !== resolvedUrlPathName || baseUrlOrigin !== resolvedUrlOrigin) {
       siteConfig.updateFetchConfig({
         overrideBaseURL: resolvedUrlOrigin,
       });