+++ CHANGELOG.md
@@ -1,3 +1,102 @@
+## [1.177.6](https://github.com/adobe/spacecat-api-service/compare/v1.177.5...v1.177.6) (2025-09-11)
+
+
+### Bug Fixes
+
+* fixing trial user create ([#1243](https://github.com/adobe/spacecat-api-service/issues/1243)) ([f3e1d4d](https://github.com/adobe/spacecat-api-service/commit/f3e1d4d1b7d9eb5bf9b7baaeb3119e87ff3bffa9))
+
+## [1.177.5](https://github.com/adobe/spacecat-api-service/compare/v1.177.4...v1.177.5) (2025-09-11)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-rum-api-client to v2.37.3 ([#1241](https://github.com/adobe/spacecat-api-service/issues/1241)) ([813ecfe](https://github.com/adobe/spacecat-api-service/commit/813ecfe4165d652aff196136c866f13fd990590f))
+
+## [1.177.4](https://github.com/adobe/spacecat-api-service/compare/v1.177.3...v1.177.4) (2025-09-10)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.62.0 ([#1240](https://github.com/adobe/spacecat-api-service/issues/1240)) ([95260d9](https://github.com/adobe/spacecat-api-service/commit/95260d901cfa002420b1b0a5fffcb803dee068b3))
+
+## [1.177.3](https://github.com/adobe/spacecat-api-service/compare/v1.177.2...v1.177.3) (2025-09-10)
+
+
+### Bug Fixes
+
+* updating lorem-ipsum check name in preflight ([#1237](https://github.com/adobe/spacecat-api-service/issues/1237)) ([c957382](https://github.com/adobe/spacecat-api-service/commit/c957382475df8e3201172edf548b954935ecc2b2))
+
+## [1.177.2](https://github.com/adobe/spacecat-api-service/compare/v1.177.1...v1.177.2) (2025-09-10)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-utils to v1.50.4 ([#1239](https://github.com/adobe/spacecat-api-service/issues/1239)) ([fe32310](https://github.com/adobe/spacecat-api-service/commit/fe323102efa150508db748bacdfcc7c5fd2ef9ad))
+
+## [1.177.1](https://github.com/adobe/spacecat-api-service/compare/v1.177.0...v1.177.1) (2025-09-09)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#1238](https://github.com/adobe/spacecat-api-service/issues/1238)) ([303278d](https://github.com/adobe/spacecat-api-service/commit/303278dd7fc6cc03a17cd4718f9813fa22828e9a))
+
+# [1.177.0](https://github.com/adobe/spacecat-api-service/compare/v1.176.13...v1.177.0) (2025-09-08)
+
+
+### Features
+
+* Llmo endpoint to fetch global files ([#1234](https://github.com/adobe/spacecat-api-service/issues/1234)) ([ad4d6a3](https://github.com/adobe/spacecat-api-service/commit/ad4d6a36b5b663605494a14de41d3c8d28ea2bd5))
+
+## [1.176.13](https://github.com/adobe/spacecat-api-service/compare/v1.176.12...v1.176.13) (2025-09-08)
+
+
+### Bug Fixes
+
+* add first name and last name in first login ([#1233](https://github.com/adobe/spacecat-api-service/issues/1233)) ([09bf0e7](https://github.com/adobe/spacecat-api-service/commit/09bf0e7676db7d05c00ca17fd02d4e48aa6ded16))
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.61.3 ([#1232](https://github.com/adobe/spacecat-api-service/issues/1232)) ([966e53e](https://github.com/adobe/spacecat-api-service/commit/966e53e1c0d2eddc57aa8154c64c513e128a8084))
+
+## [1.176.12](https://github.com/adobe/spacecat-api-service/compare/v1.176.11...v1.176.12) (2025-09-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-utils to v1.50.2 ([#1231](https://github.com/adobe/spacecat-api-service/issues/1231)) ([419990b](https://github.com/adobe/spacecat-api-service/commit/419990b4ddf6766e4a0e534665b06d78f11f1736))
+
+## [1.176.11](https://github.com/adobe/spacecat-api-service/compare/v1.176.10...v1.176.11) (2025-09-06)
+
+
+### Bug Fixes
+
+* adding variationName to auto fix sqs message ([#1128](https://github.com/adobe/spacecat-api-service/issues/1128)) ([ab7c837](https://github.com/adobe/spacecat-api-service/commit/ab7c837aaa1f92a3c7583cd60c19fe7565ce3085))
+
+## [1.176.10](https://github.com/adobe/spacecat-api-service/compare/v1.176.9...v1.176.10) (2025-09-06)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.61.2 ([#1229](https://github.com/adobe/spacecat-api-service/issues/1229)) ([c6f1267](https://github.com/adobe/spacecat-api-service/commit/c6f1267057becceeef76690352f6f5bf8c1fd91d))
+
+## [1.176.9](https://github.com/adobe/spacecat-api-service/compare/v1.176.8...v1.176.9) (2025-09-06)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1228](https://github.com/adobe/spacecat-api-service/issues/1228)) ([2116420](https://github.com/adobe/spacecat-api-service/commit/2116420fb6f38cfba9c49b8490576feaafd7ff2c))
+
+## [1.176.8](https://github.com/adobe/spacecat-api-service/compare/v1.176.7...v1.176.8) (2025-09-06)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#1227](https://github.com/adobe/spacecat-api-service/issues/1227)) ([76c633f](https://github.com/adobe/spacecat-api-service/commit/76c633fce4fe33ef9f29171fdd06c866e7ec12f4))
+
+## [1.176.7](https://github.com/adobe/spacecat-api-service/compare/v1.176.6...v1.176.7) (2025-09-05)
+
+
+### Bug Fixes
+
+* onboard issues ([#1222](https://github.com/adobe/spacecat-api-service/issues/1222)) ([96f976f](https://github.com/adobe/spacecat-api-service/commit/96f976f361a93b90d1ace1a0f74520b21bc520fe))
+
 ## [1.176.6](https://github.com/adobe/spacecat-api-service/compare/v1.176.5...v1.176.6) (2025-09-05)